mips64 progress

+++ crates/kernel/src/types.rs
@@ -16,6 +16,6 @@ pub struct Proof {
     pub step_input: Vec<u8>,
     pub oracle_key: Option<Vec<u8>>,
     pub oracle_value: Option<Vec<u8>>,
-    pub oracle_offset: Option<u32>,
+    pub oracle_offset: Option<u64>,
     pub oracle_input: Option<Vec<u8>>,
 }
